installs TableGen and uses it to generate diagnostics
TableGen is an effective way to organise and generate things such as
diagnostic data and recognised symbols. It's a little bit Clangey at
the moment, but that can probably be smoothed out over time, but this
is to be expected since the structure and a lot of the code is directly
imported from Clang.