adds some of Clang's TableGen logic so we can generate part of the lexer
We're going to use TableGen because it's a convenient way to organise
code for lexing. For more info about TableGen, please see
https://llvm.org/docs/TableGen/.

Most of the C++ in this commit is from Clang and shouldn't be reviewed.
The code worth reviewing is in `SchreiberCommentCommandInfoEmitter.cpp`,
and that's fairly bare-bones at the moment.

Unlike the rest of the project, the TableGen component does not have
unit tests because it is generating code for the project, and so its
output will always be tested by building Schreiber.